refactor: 캡슐 컴포넌트 분리
